Try to estimate the calibration matrix and the offset.
Researching literature and least square regularization methods I saw that using and augmented vector to include the offset in the calculations on methods with regularization may give solutions with priority to cross the center, (not estimating well the bias). So instead using the difference between the mean and the actualv value for the reference and the raw data allows to obtain the calibration matrix without bias and then offset can be estimated.